moved all servos
