+++ include/servos.hpp
@@ -1,6 +1,31 @@
 #pragma once
 
+#include <map>
 #include <mbed.h>
+#include <std_msgs/msg/u_int32_multi_array.h>
+
+static std::map<double, uint8_t> servo_voltage_configuration{
+    {5.0, 0},
+    {6.0, 1},
+    {7.4, 2},
+    {8.6, 3}
+};
+
+enum Servos{
+    servo0,
+    servo1,
+    servo2,
+    servo3,
+    servo4,
+    servo5,
+    SERVOS_COUNT
+};
+
+extern std_msgs__msg__UInt32MultiArray servos_command_msg;
+
+void init_servos();
+void fill_servos_command_msg(std_msgs__msg__UInt32MultiArray *msg);
+void servos_command_callback(const void* message);
 
 namespace rosbot_sensors {
 
@@ -41,3 +66,5 @@ private:
 };
 
 } // namespace rosbot_sensors
+
+extern rosbot_sensors::ServoManger servo_manager;
